// src/server/schema.d.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
  "/pets": {
    get: operations["listPets"];
    post: operations["createPets"];
  };
  "/pets/{petId}": {
    get: operations["showPetById"];
  };
}

export interface components {
  schemas: {
    Pet: {
      /** Format: int64 */
      id: number;
      name: string;
      tag?: string;
    };
    Pets: components["schemas"]["Pet"][];
    Error: {
      /** Format: int32 */
      code: number;
      message: string;
    };
  };
}

export interface operations {
  listPets: {
    parameters: {
      query: {
        /** How many items to return at one time (max 100) */
        limit?: number;
      };
    };
    responses: {
      /** A paged array of pets */
      200: {
        headers: {
          /** A link to the next page of responses */
          "x-next"?: string;
        };
        content: {
          "application/json": components["schemas"]["Pets"];
        };
      };
      /** unexpected error */
      default: {
        content: {
          "application/json": components["schemas"]["Error"];
        };
      };
    };
  };
  createPets: {
    responses: {
      /** Null response */
      201: unknown;
      /** unexpected error */
      default: {
        content: {
          "application/json": components["schemas"]["Error"];
        };
      };
    };
  };
  showPetById: {
    parameters: {
      path: {
        /** The id of the pet to retrieve */
        petId: string;
      };
    };
    responses: {
      /** Expected response to a valid request */
      200: {
        content: {
          "application/json": components["schemas"]["Pet"];
        };
      };
      /** unexpected error */
      default: {
        content: {
          "application/json": components["schemas"]["Error"];
        };
      };
    };
  };
}

export interface external {}